adds aperture photometry operation and utils
